move new version of functions to separate file
