+++ tracer/src/Datadog.Trace.ClrProfiler.Native/Datadog.Trace.ClrProfiler.Native.vcxproj.filters
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
+    <ClCompile Include="always_on_profiler_clr_helpers.h" />
     <ClCompile Include="always_on_profiler.cpp" />
     <ClCompile Include="calltarget_tokens.cpp" />
     <ClCompile Include="class_factory.cpp" />
@@ -41,6 +42,7 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="always_on_profiler_clr_helpers.h" />
     <ClInclude Include="always_on_profiler.h" />
     <ClInclude Include="calltarget_tokens.h" />
     <ClInclude Include="class_factory.h" />
